requirements.txt
telethon~=1.39.0
prompt~=0.4.1
rich~=13.7.1
inquirerpy~=0.3.4
typing
pytz~=2024.1
pyTelegramBotAPI~=4.14.0
schedule~=1.2.0
tinydb~=4.8.0
simplejson~=3.19.2
python-dotenv~=1.0.0
flask~=3.0.0
